requirements.txt
google-genai>=0.2.0
requests>=2.31.0
beautifulsoup4>=4.12.0
python-dotenv>=1.0.0
praw>=7.7.0
tavily-python>=0.3.0
lxml>=5.0.0
rich>=13.0.0
